+++ src/main/java/group11/Hockey/db/IPlayerDb.java
@@ -2,5 +2,5 @@ package group11.Hockey.db;
 
 public interface IPlayerDb {
 	public boolean insertLeagueFreeAgents(String leagueName, String freeAgentName, String position, float playerSkating,
-			float playerShooting, float playerChecking, float playerSaving, int age);
+			float playerShooting, float playerChecking, float playerSaving, float age);
 }
